refactor(web-domains): Profile Image -> Avatar로 일괄 변경 및 objectFit: cover 적용 (#157)
* refactor: home 지면, 릴레이 질문 답변 관련 지면 objectFit cover 일괄 적용

* refactor: 릴레이 질문 시작 지면 objectFit cover 적용

* refactor: 나머지 지면 프로필 관련 Image 태그는 Avatar 로 일괄 변경 및 ObjectFit 적용

* rename: Avatar web-domains common 으로 이동

* refactor: Avatar 컴포넌트 css 대신 style 사용토록 변경
