+++ packages/web-domains/src/about-me/features/components/Profile/Profile.tsx
@@ -2,8 +2,9 @@
 
 import { Badge, Txt } from '@sambad/sds/components';
 import { borderRadiusVariants, colors } from '@sambad/sds/theme';
-import Image from 'next/image';
 import { Fragment } from 'react';
+
+import { Avatar } from '@/common/components/Avatar/Avatar';
 
 import { badgeContainerCss, nameCss } from './styles';
 import { generateAge, generateGender } from './utils';
@@ -31,8 +32,8 @@ export const Profile = (props: ProfileProps) => {
   return (
     <Fragment>
       {imageUrl != null && (
-        <Image
-          src={imageUrl}
+        <Avatar
+          imageUrl={imageUrl}
           width={80}
           height={80}
           alt={`${name}_프로필_이미지`}
