+++ packages/web-domains/src/result/features/components/SelectedDataCard/Profile.tsx
@@ -1,8 +1,9 @@
 'use client';
 
 import { Txt, TxtProps } from '@sambad/sds/components';
-import Image from 'next/image';
 import { HTMLAttributes } from 'react';
+
+import { Avatar } from '@/common/components/Avatar/Avatar';
 
 import { profileCss } from './styles';
 
@@ -18,7 +19,7 @@ export const Profile = (props: ProfileProps) => {
 
   return (
     <div css={profileCss} {...restProps}>
-      {imgUrl && <Image src={imgUrl} width={imgSize} height={imgSize} alt={`${name}_프로필_이미지`} />}
+      {imgUrl && <Avatar imageUrl={imgUrl} width={imgSize} height={imgSize} alt={`${name}_프로필_이미지`} />}
       {name && <Txt typography={nameTypography}>{name}</Txt>}
     </div>
   );
